bot_left_channel + refactoring
